test(rpz): dissolve coverage_gaps_test.go into the themed test files
The file was named for the process that produced it, not for what it
pins — meaningless to a future reader. The feed lifecycle test joins
axfr_test.go; the response-action, neutral-gate, and decoded-path tests
join sidecar_test.go with the serve helper they share. No test changed.
